requirements.txt
pandas
pyarrow
duckdb
pymongo[srv]
python-dotenv
scikit-learn
jupyter
